making tests more port-independent and applying a little bit of DRY
